Move clang-native handling into tests directory
This stuff is only used by tests.

Also, some cleanup of test_benchmark.py
